Fix bug on the 'if' statement that adds 'active' class
I find bug on Lektor.utils.is_path_child_of(). In case that
this.page_num is not None and href starts with '/@', the function
returns True. In my case, this._path is '/blog',
this.page_num is 1 (in other words, this.path is '/blog@1')
and href=='/@tag/'.
I may fix bug another way later. For instance, forks from Lektor
or creates PR to Lektor